replacing TCA Interface /w wire.h
