+++ platformio.ini
@@ -82,6 +82,7 @@ lib_deps =
 	paulstoffregen/OneWire@^2.3.7
 	adafruit/Adafruit BusIO@^1.16.1
 	adafruit/Adafruit GFX Library@^1.11.9
+	https://github.com/youseetoo/TCA9555.git
 build_flags = 
 	${env:esp32dev.build_flags}
 	-DCORE_DEBUG_LEVEL=5
